Introduce `reaver::bound_configuration`.
The invariant for `bound_configuration<Args...>` is that only tags that
are present in `Args...` are allowed, and that they are always present.
Not all parts of the code committed here are tested (for example, the
conditions disabling the public constructors are not tested in any way);
more tests will follow, but right now the most important features are
tested.