Add support for default values; add support for `parsed_type`.
`parsed_type` is meant as a way to specify a different type that is to
be passed to `boost::program_options`. This way, if a tag specifies,
say, an `enum` as its `type`, and specifies a `construct` from string to
that enum, and does `using parsed_type = std::string;`, then it just
works - program_options thinks it's parsing a string, while the final
value that ends up in the configuration becomes the enum specified with
`type`.