function: introduce an std::function replacement.
Still not a fully operational thing. Will need additional checks on
construction and better type error handling, but currently it's
introduced because std::function is not nearly good enough for
reaver::variant, hence it's not good enough for future (namely,
stdlibc++'s move operations of std::function are not noexcept, which is
a terrible PITA). Hopefully I can get this fixed soon. (A support, with
a secondary storage area, for types that aren't noexcept move
constructible in reaver::variant is planned, but not focused on.)